requirements.txt
fastapi
uvicorn[standard]
httpx
gunicorn
pdfplumber
python-docx
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
asyncpg
